pyproject.toml
[project]
name = "digest"
version = "0.1.0"
description = "Weekly article digest bot: RSS → LLM filtering → Telegram"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.0",
    "anthropic>=0.40",
    "beautifulsoup4>=4.14",
    "feedparser>=6.0",
    "google-genai>=1.0",
    "httpx>=0.28",
    "lxml>=6.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
]
